shit do chat

+++ src/core/threads/C_tLeaveRoomAccess.h
@@ -11,9 +11,9 @@ class C_tLeaveRoomAccess : public C_Thread {
 private:
     C_Monitor& m_monitorrfid;
     C_Monitor& m_monitorservoroom;
-    C_RDM6300& m_rfidExit;       // Sensor montado no interior para sair
+    C_RDM6300& m_rfidExit;       
     C_Mqueue& m_mqToDatabase;
-    C_Mqueue& m_mqToLeaveRoom;   // Fila de resposta específica para a saída
+    C_Mqueue& m_mqToLeaveRoom;   
     C_Mqueue& m_mqToActuator;
 
     int m_failedAttempts;
@@ -28,8 +28,10 @@ public:
 
     virtual ~C_tLeaveRoomAccess();
 
-    void generateDescription(uint8_t userId, char* buffer, size_t size);
-    void sendLog(uint8_t userId, uint16_t accessLevel);
+    // void generateDescription(uint8_t userId, char* buffer, size_t size);
+    void generateDescription(uint32_t userId, char* buffer, size_t size);
+    // void sendLog(uint8_t userId, uint16_t accessLevel);
+    void sendLog(uint32_t userId, uint32_t accessLevel);
     void run() override;
 };
